Turn on -Wsuggest-attribute=(pure|const|cold)
Also: fix all of the warnings that doing so introduced
Also: Implement a couple of missing wchar-related functions from <stdlib.h> and <wchar.h>

+++ kos/src/libregdump/arch/i386/x86.c
@@ -319,7 +319,7 @@ libregdump_do_segment(struct regdump_printer *__restrict self,
 	if (value & 4) {
 		if (!*pldt)
 			*pldt = __sldt();
-		if ((*pldt & ~7) >= tab.dt_limit + 1) {
+		if ((uint16_t)(*pldt & ~7) >= (uint16_t)(tab.dt_limit + 1)) {
 			PRINT("] ");
 			format(REGDUMP_FORMAT_ERROR_PREFIX);
 			printf("CORRUPT LDT: %%ldt(%#.2x) >= %%gdt.size(%#.2x)%s\n",
